requirements.txt
Pillow==10.4.0
pytesseract==0.3.13
streamlit==1.38.0
regex